+++ Source/ACE/Entity/Door.cs
@@ -1,4 +1,4 @@
-﻿using System.Collections.Generic;
+using System.Collections.Generic;
 using ACE.Entity.Enum.Properties;
 using ACE.Entity.Enum;
 using ACE.Entity.Actions;
@@ -35,10 +35,13 @@ namespace ACE.Entity
         public Door(AceObject aceO)
             : base(aceO)
         {
-            // Set PhysicsState defaults.. Leaving commented for now to read in what was pcapped
-            // PhysicsState = 0;
-            // ReportCollision = true;
-            // HasPhysicsBsp = true;
+            Door = true;
+            Stuck = true; Attackable = true;
+            
+            SetObjectDescriptionBools();
+
+            // This likely will change to be based on reading a dat file to determine if this exists...
+            HasPhysicsBsp = true;
 
             if (!DefaultOpen)
             {